Added notes (comments) & enabled redraw for MacOS
It turns out that the strategy of temporarily adjusting the 'shellpipe'
option is not achieving the desired result when on MacOS with iTerm.
Specifically, it does not prevent :grep from clearing the screen like
:make does (I think it is still switching into "cooked" mode)

Therefore, I have enabled the `redraw!` command when on MacOS.

However, it still does not solve the problem, but this commit includes
notes via comments I do not want to lose.