Fixing bug when the path has a trailing slash
This arises from the change we want to keep:  passing the selected path
into the :grep command instead of :cd'ing first and restoring CWD after.
This has the benefit of the quickfix title containing the path that was
searched -- really helpful when traversing the multiple quickfix lists.

It may be that surrounding the path with quotes is not the right thing
to do for handling paths with spaces.   Perhaps it should be that we use
`fnameescape()` to escape everything?   Or perhaps we just escape the
spaces themselves?

For now, since surrounding the path with spaces handles most of what I
encounter - (on windows & macos) - I will continue that, and just fix
the edge case where the path has a final trailing backslash.

This of course only happens on windows, and I *think* it only happens
when searching the "root" of a drive.  A message will highlight when
this assumption is wrong, with the search still completing hopefully.